chore(fix-dep): Add 'graal-sdk' dependency to provide classes for substitutions (no longer transitive since Quarkus 3.3)
In line with the Quarkus migration guide, we now consistently rely on the 'org.graalvm.sdk:graal-sdk' artifact for GraalVM support. The dependencies to this artifact are marked as provided in Quarkus, making them non-transitive. If you want to include GraalVM substitutions in your applications, please add it as a dependency yourself from now on. It is part of the BOM, so you don’t have to define the version.

Signed-off-by: Sun Seng David TAN <sutan@redhat.com>